requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
charset-normalizer==3.3.2
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
gast==0.3.3
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==2.10.0
idna==3.10
importlib-metadata==6.7.0
Jinja2==3.1.4
joblib==1.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.1.0
mne==1.3.1
numpy==1.18.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
platformdirs==4.0.0
pooch==1.8.2
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==5.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.21.2
scipy==1.7.3
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.1
PyYAML==5.1.1